Updated to godot beta 9
This was originally to port to android, but it seems completely unfeasible for many reasons... For now though, this should make it easier to transition to a stable godot 4 version in the future. The documentation will have to be updated, so older commits are probably best to use for now.