Huge performance increase
Before, sand slide randomly sampled particles to move in the hopes that large enough iterations would be fast enough to mimic real particles. However, this was extremely slow; now, the engine goes through each particle on every frame. Some more tricks are required to keep a similar amount of noise as the old engine, but the performance is too big to pass up.