Refactored code for element visuals
This is a massive overhaul from the 500 line shader script. Now, element visuals are defined using resources. This allows the shader to rely on branching MUCH less, making it possible to render on phones! An android build is in the works, but for now all of the colors will have to be readded. The new gradient feature may not be feasible however, so for now all elements will have the basic gradient from before. The sky and tint is also disabled until the android build is stable.